QC fields feature (#126)
This PR adds a new QC fields feature: allow a field to be duplicated in the `field_to_zap.xml` file for a QC on the GUI. QC fields are not integrated in the final file. QC fields must begin with a `qc-` prefix in their `name` tag in the `field_to_zap.xml` file. E.g. 

```xml
<!-- not qc field -->
<item>
  <name>0008,0022</name>
  <description>AcquisitionDate</description>
  <editable>yes</editable>
</item>

<!-- qc field, associated with the above one -->
<item>
  <name>qc-0008,0022</name>
  <description>AcquisitionDate</description>
  <editable>yes</editable>
</item>
```

Also, fixes small issues on path linking.